requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
